chore:refactor: config and apply eslint rules (#63)
* add eslint deps and rules

* eslint --fix

* chore: lint (remove unused)

* feat: on board interep with gh (#61)

* refactor: extract interepOnboarding views in separate files

* define AuthGithub / AuthTwitter components

* render joinGroupView with GitHub reputation params

* can reset/logout
