+++ .eslintignore
@@ -1,0 +1,3 @@
+.idea
+build
+webpack.ui.config.js
