+++ package.json
@@ -12,6 +12,8 @@
     "test:coverage": "NODE_ENV=test jest --coverage",
     "format:check": "prettier --check 'src'",
     "format:fix": "prettier --write 'src'",
+    "lint:check": "eslint . --ext .ts --ext .tsx",
+    "lint:fix": "eslint . --ext .ts --ext .tsx --fix",
     "prepare": "husky install"
   },
   "dependencies": {
@@ -78,12 +80,19 @@
     "@types/redux-logger": "^3.0.8",
     "@types/sinon": "^10.0.11",
     "@types/webtorrent": "^0.109.3",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
     "browserify": "^17.0.0",
     "concurrently": "^5.1.0",
     "constants-browserify": "^1.0.0",
     "copy-webpack-plugin": "^5.0.5",
     "css-loader": "^3.2.1",
     "eslint": "^5.16.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^3.5.2",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.5.1",
+    "eslint-plugin-promise": "^6.1.1",
     "fake-indexeddb": "^3.1.7",
     "file-loader": "^5.0.2",
     "html-webpack-plugin": "~5.3.2",
