+++ src/components/ChatMenu/index.tsx
@@ -29,7 +29,6 @@ import { useThemeContext } from '../ThemeContext';
 export default function ChatMenu(): ReactElement {
   const selected = useSelectedLocalId();
   const selecteduser = useUser(selected?.address);
-  const history = useHistory();
   const dispatch = useDispatch();
   const chatIds = useChatIds();
   const [showingCreateChat, setShowingCreateChat] = useState(false);
@@ -109,7 +108,7 @@ export default function ChatMenu(): ReactElement {
           <Icon className="text-gray-400 mx-2" fa="fas fa-search" size={0.75} />
         </Input>
       </div>
-      {!!showingCreateChat &&
+      {showingCreateChat &&
         (searchResults?.length ? (
           searchResults.map(chat => (
             <ChatMenuItem
