+++ src/components/Web3Button/index.tsx
@@ -60,7 +60,7 @@ export default function Web3Button(props: Props): ReactElement {
         const ens = await fetchNameByAddress(account);
         setEnsName(ens);
       } else {
-        let id = selectedLocalId || identities[0];
+        const id = selectedLocalId || identities[0];
 
         if (id?.type !== 'zkpr_interrep' && id?.type !== 'taz') {
           setEnsName('');
@@ -72,7 +72,7 @@ export default function Web3Button(props: Props): ReactElement {
   }, [account, identities, selectedLocalId]);
 
   let btnContent;
-  let id = selectedLocalId || identities[0];
+  const id = selectedLocalId || identities[0];
 
   const onClick = useCallback(() => {
     if (!id) {
@@ -223,7 +223,7 @@ function UserMenuable(props: {
     });
   }, []);
 
-  let items: ItemProps[] = [];
+  const items: ItemProps[] = [];
 
   if (selectedLocalId || identities.length) {
     items.push({
@@ -386,7 +386,7 @@ function CurrentUserItem(props: {
 
   const gotoProfile = useCallback(() => {
     if (!selectedUser) return;
-    const { ens, name, address } = selectedUser;
+    const { ens, address } = selectedUser;
     history.push(`/${ens || address}`);
     props.closePopup();
   }, [selectedUser]);
