+++ src/ducks/posts.test.ts
@@ -11,7 +11,6 @@ const {
     fetchHomeFeed,
     fetchLikedBy,
     fetchRepliedBy,
-    setPost,
     setLiked,
     setReposted,
     setBlockedPost,
@@ -20,7 +19,6 @@ const {
     incrementLike,
     decrementRepost,
     decrementLike,
-    unsetPost,
   },
 } = ducks;
 
@@ -31,8 +29,8 @@ describe('Posts Duck', () => {
 
   it('should fetch meta', async () => {
     const messageId = '0xmeta/0000000000000000000000000000000000000000000000000000000000000000';
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: {
@@ -57,8 +55,8 @@ describe('Posts Duck', () => {
     const messageId = 'e6789d8ea65b57efd365ada389924246e4bd7be9c109a7fe294646831f67db8b';
     gunStub.get
       .withArgs('message/e6789d8ea65b57efd365ada389924246e4bd7be9c109a7fe294646831f67db8b')
-      // @ts-ignore
       .returns(
+        // @ts-ignore
         Promise.resolve({
           type: 'POST',
           subtype: 'REPLY',
@@ -68,11 +66,11 @@ describe('Posts Duck', () => {
           },
         })
       )
-      // @ts-ignore
       .withArgs('payload')
+      // @ts-ignore
       .returns(Promise.resolve({ content: 'fetch post' }));
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: {
@@ -88,8 +86,8 @@ describe('Posts Duck', () => {
   });
 
   it('should fetch posts', async () => {
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: [
@@ -127,8 +125,8 @@ describe('Posts Duck', () => {
   });
 
   it('should fetch liked by', async () => {
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: [
@@ -167,8 +165,8 @@ describe('Posts Duck', () => {
   });
 
   it('should fetch replied by', async () => {
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: [
@@ -207,8 +205,8 @@ describe('Posts Duck', () => {
   });
 
   it('should fetch homefed', async () => {
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: [
@@ -247,8 +245,8 @@ describe('Posts Duck', () => {
   });
 
   it('should fetch tagfeed', async () => {
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: [
@@ -287,8 +285,8 @@ describe('Posts Duck', () => {
   });
 
   it('should fetch replies', async () => {
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: [
