+++ src/ducks/zkpr.test.ts
@@ -1,13 +1,13 @@
 import { ducks, fetchStub, postWorkMessageStub, store, zkprStub } from '~/testUtils';
 
 const {
-  zkpr: { connectZKPR, createZKPRIdentity, maybeSetZKPRIdentity },
+  zkpr: { connectZKPR },
 } = ducks;
 
 describe('ZKPR duck', () => {
   it('should connect to zkpr', async () => {
-    // @ts-ignore
     fetchStub.returns(
+      // @ts-ignore
       Promise.resolve({
         json: async () => ({
           payload: {
