+++ src/pages/SettingView/index.tsx
@@ -1,13 +1,9 @@
-import React, { ReactElement, ReactNode, useCallback, useEffect, useState } from 'react';
+import React, { ReactElement, ReactNode, useCallback, useState } from 'react';
 import classNames from 'classnames';
 import './setting.scss';
 import { Redirect, Route, Switch, useHistory, useLocation } from 'react-router';
 import Icon from '@components/Icon';
-import Avatar from '@components/Avatar';
 import { useDispatch } from 'react-redux';
-import { fetchAddressByName, useUser } from '@ducks/users';
-import Web3 from 'web3';
-import { getHandle, getName } from '~/user';
 import SwitchButton from '@components/SwitchButton';
 import { setTheme, useSetting } from '@ducks/app';
 
@@ -118,39 +114,6 @@ function DisplaySetting(): ReactElement {
         </select>
       </SettingRow>
     </>
-  );
-}
-
-function UserRow(props: { name: string }): ReactElement {
-  const history = useHistory();
-  const [username, setUsername] = useState('');
-
-  const dispatch = useDispatch();
-  const user = useUser(username);
-
-  useEffect(() => {
-    (async () => {
-      if (!Web3.utils.isAddress(props.name)) {
-        const address: any = await dispatch(fetchAddressByName(props.name));
-        setUsername(address);
-      } else {
-        setUsername(props.name);
-      }
-    })();
-  }, [props.name]);
-
-  if (!user) return <></>;
-
-  return (
-    <div
-      className="flex flex-row flex-nowrap px-3 py-2 cursor-pointer hover:bg-gray-50 items-center"
-      onClick={() => history.push(`/${user.ens || user.address}/`)}>
-      <Avatar address={user.address} className="w-10 h-10 mr-3" />
-      <div className="flex flex-col flex-nowrap justify-center">
-        <div className="font-bold text-light hover:underline">{getName(user, 8, 6)}</div>
-        <div className="text-xs text-gray-500">@{getHandle(user, 8, 6)}</div>
-      </div>
-    </div>
   );
 }
