+++ src/util/merkle.ts
@@ -1,14 +1,6 @@
 import config from './config';
 import { MerkleProof } from '@zk-kit/protocols';
 
-type PathData = {
-  path_elements: string[] | string[][];
-  path_index: number[];
-  root: string;
-};
-
-export type ValidGroups = 'zksocial_all' | 'interep_twitter_unrated';
-
 export const findProof = async (
   group: string,
   idCommitment: string,
