+++ src/util/sse.ts
@@ -96,8 +96,7 @@ class SSE extends EventEmitter2 {
         topics,
       }),
     });
-    const json = await resp.json();
-    return json;
+    return await resp.json();
   };
 }
